:lipstick: feat: rfid para Long

+++ BackEnd/src/main/java/com/MapView/BackEnd/dtos/Equipment/EquipmentCreateDTO.java
@@ -10,7 +10,7 @@ public record EquipmentCreateDTO(
         String name_equipment,
         @Positive
         @Size(min = 15)
-        int rfid,
+        Long rfid,
         @NotBlank(message = "Type cannot be blank")
         String type,
         EnumModelEquipment model,
